Address map header addd to the proyect
The goal of the header is  set the address using only one file and avoid change
multiple files if the address map needs adjustments.

address_map.vh created to parameterize the modules.
the address map it is for the new architecture proposed to implement.
the new feautres of the board (channel_configuration, multiple PPS dividers,
send data to TIVA by SPI).