Creation of mux data
Mux to control the read data send to the SPI to be transmited
in the MISO port.

This mux has as inputs the data read of each module (PPS_DIVIER and PULSE_GEN) also
the data read of the main_memory.

The output of the mux is update each clock cycle (rising edge)
To avoid problems with pure combinational multiplexer.